task resolved

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -9,7 +9,7 @@
 */
 void largest_prime_factor(long int num)
 {
-	int primn, largest;
+	long int  primn, largest;
 
 	while (num % 2 == 0)
 		num /= 2;
